+++ src/march/march.h
@@ -6,7 +6,7 @@ extern "C" {
 #endif
 
 enum { MARCH_NTRI = 48 };
-extern const double MARCH_O[][3];
+extern const int MARCH_O[][3];
 int march_cube(double u[8], /**/ int* ntri, double* tri);
 int march_cube_location(
     double u[8], /**/ int* ntri, double* tri, int* x, int* y, double* a);
